% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_info.R
\name{company_info}
\alias{company_info}
\title{Company identifier information.}
\usage{
company_info(symbol)
}
\arguments{
\item{symbol}{String of the company trading symbol in lower or upper case.}
}
\value{
Returns a data frame consisting of the input company's name
("Company"), stock symbol ("Symbol"), "CIK" - central index key, or the unique
identifier used by the SEC's EDGAR system for each company, "SIC" -
standard industrial classification, a code identifying the broader industry
within which the company operates, and the end date for the company's fiscal
year ("Fiscal_end").
}
\description{
\code{company_info} Returns company identification information used by the
SEC to organize company financial reports. The identifiers returned by this
function facilitate report extraction in \code{\link{reports}}.
}
\examples{
company_info("tsla")
company_info("AAPL")

\dontrun{
company_info(aapl)
}

}
