% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials_doAll.R
\name{financials_doAll}
\alias{financials_doAll}
\title{Wrapper for financials_sec to return panel or time series financial data}
\usage{
financials_doAll(symbols, years, report_type = "10")
}
\description{
Wrapper for financials_sec to return panel or time series financial data
}
