% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialsR-package.R
\docType{package}
\name{FinancialsR-package}
\alias{FinancialsR}
\alias{FinancialsR-package}
\title{FinancialsR: Extract and tidy SEC corporate financial data for analysis}
\description{
Connects to SEC's EDGAR database to extract financial filings
    returning a tidied dataframe for convenient analysis in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loganstundal/financialsR}
  \item Report bugs at \url{https://github.com/loganstundal/financialsR/issues}
}

}
\author{
\strong{Maintainer}: Logan Stundal \email{logan.stundal@gmail.com}

}
\keyword{internal}
