% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_technicals.R
\name{plot_technicals}
\alias{plot_technicals}
\title{Generates a plot of technical data from a company's stock.}
\usage{
plot_technicals(
  symbol,
  from = NULL,
  to = NULL,
  theme = "dark",
  typical_price = FALSE,
  return_stock_price = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{symbol}{Company stock ticker as string.}

\item{from, to}{Start and end date for stock prices as strings in format:
"YYYY-MM-DD". Note, difference between the start date and end date must be
at least 27 trading days in order to calculate the 26-day EMA. Periods
shorter than this will produce an error in which case use an earlier start
date. Default values are "2018-01-01" for "from" and today for "to.}

\item{theme}{Setting for plot theme. Options include "dark", the default, or "bright".}

\item{typical_price}{True or False. Incidator to use daily typical price
defined as "(High + Low + Close) /3; default is False which uses daily
closing price.}

\item{return_stock_price}{True or False. Indicator to return underlying
stock price data as a tidied data frame.}

\item{show_plot}{True or False. Indicator to show plot.}
}
\value{
Returns a ggplot object of stock price, EMA, MACD. Optionally returns
a data frame consisting of underlying stock price data.
}
\description{
\code{plot_technicals} Returns a plot of company stock price along
with optional data frame of underling price data.
}
\examples{
plot_technicals("TSLA"
plot_technicals("TSLA", from = as.character(today() - 150))
}
