% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{Retrieve corporate filings from SEC}
\usage{
reports(
  company,
  type = "any",
  prior_to = NULL,
  exclude_amended = FALSE,
  limit_results = 20
)
}
\arguments{
\item{company}{Company stock ticker entered as a string (e.g., "tsla")}

\item{type}{Report filing type from the SEC. Options include the following:
10-K, 10-Q, 10-K/A, 10-Q/A, 10, 8-K, any.}

\item{prior_to}{Prior to exclusion date entered as a string in the following
format: YYYYMMDD}

\item{exclude_amended}{Boolean to excluded amended filing results.}

\item{limit_results}{Numeric value indicating the limit on returned results.
Must be one of: 10, 20, 40, 80, or 100.}
}
\value{

}
\description{
\code{reports} does xyz
}
\examples{
reports("tsla")
}
